fix: close transcript projection integration gaps
