+++ src/config/sessions/session-transcript-projection-rebuild.test.ts
@@ -32,12 +32,12 @@ import {
   projectionRow as row,
   readDisplayRowIdentities,
   readDisplaySnapshot,
+  readRequiredSourceGeneration,
   serializeDisplayTables,
 } from "./session-transcript-display.test-support.js";
-import type { SessionTranscriptProjectionSourceRow } from "./session-transcript-projection-rebuild.js";
 import { waitForSessionTranscriptIndexReconcile } from "./session-transcript-reconcile.js";
-import { readSessionTranscriptSourceGenerationTokenInTransaction } from "./session-transcript-source-generation.js";
 
+type SessionTranscriptProjectionSourceRow = ReturnType<typeof row>;
 const SESSION_ID = "projection-session";
 const tempDirs = useAutoCleanupTempDirTracker(afterEach);
 
@@ -106,18 +106,11 @@ describe("canonical session transcript projection", () => {
               "INSERT INTO transcript_events (session_id, seq, event_json, created_at) VALUES (?, ?, ?, ?)",
             )
             .run(sessionId, seq, JSON.stringify(event), seq + 1);
-          const sourceGeneration = readSessionTranscriptSourceGenerationTokenInTransaction(
-            database.db,
-            sessionId,
-          );
-          if (!sourceGeneration) {
-            throw new Error("expected transcript source generation");
-          }
           appendEligibleSessionTranscriptDisplayRowInTransaction(database.db, {
             event,
             seq,
             sessionId,
-            sourceGeneration,
+            sourceGeneration: readRequiredSourceGeneration(database.db, sessionId),
           });
         },
         { agentId: scope.agentId, env },
